Fix IDCommentText regexp
The regexp which extracts comment text was picking up IDComment-CommentText
divs which had no valid ID or content. This made Python go a little skatty on
Precocious comic ID#820, which has... you guessed it, empty comments like this.

Regex changed to require one or more numbers, not zero or more (this was a
stupid bug anyway)